modificacion y nuevos archivos

+++ Aereos.java
@@ -1,6 +1,6 @@
 public abstract class Aereos extends Vehiculo{
     
-    public Aereos(String marca, String modelo, int velocidad, int capacidad_carga){
+    public Aereos(String marca, String modelo, double velocidad, double capacidad_carga){
         super(marca, modelo, velocidad, capacidad_carga);
         
     }
