+++ Terrestres.java
@@ -1,6 +1,6 @@
 public abstract class Terrestres extends Vehiculo{
     
-    public Terrestres(String marca, String modelo, int velocidad, int capacidad_carga){
+    public Terrestres(String marca, String modelo, double velocidad, double capacidad_carga){
         super(marca, modelo, velocidad, capacidad_carga);
         
     }
